Adde annotations to the new orms

+++ src/main/java/com/jazzybruno/example/v1/models/Course.java
@@ -16,7 +16,7 @@ public class Course {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column
-    private int courseId;
+    private Long courseId;
     @NotNull
     @Column
     private String name;
